--- get-blog2.go
@@ -1,52 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"io/ioutil"
-	"log"
-	"net/http"
-	"os"
-
-	"github.com/itchyny/gojq"
-)
-
-func curl() []byte {
-	DEVAPIKEY := os.Getenv("DEVAPIKEY")
-	client := &http.Client{}
-	req, err := http.NewRequest("GET", "https://dev.to/api/articles/me/unpublished", nil)
-	if err != nil {
-		log.Fatal(err)
-	}
-	req.Header.Set("api-key", DEVAPIKEY)
-	resp, err := client.Do(req)
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer resp.Body.Close()
-	bodyText, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		log.Fatal(err)
-	}
-	//return fmt.Printf("%s\n", bodyText)
-	return bodyText
-}
-
-func main() {
-	// Parse JSON
-	query, err := gojq.Parse(".")
-	if err != nil {
-		log.Fatalln(err)
-	}
-	input := curl()
-	iter := query.Run(input) // or query.RunWithContext
-	for {
-		v, ok := iter.Next()
-		if !ok {
-			break
-		}
-		if err, ok := v.(error); ok {
-			log.Fatalln(err)
-		}
-		fmt.Printf("%#v\n", v)
-	}
-}
