+++ src/get-blog-id.go
@@ -1,0 +1,52 @@
+package main
+
+import (
+	"encoding/json"
+	"fmt"
+	"log"
+	"net/http"
+	"os"
+
+	"github.com/itchyny/gojq"
+)
+
+func curl() interface{} {
+	DEVAPIKEY := os.Getenv("DEVAPIKEY")
+	client := &http.Client{}
+	req, err := http.NewRequest("GET", "https://dev.to/api/articles/me/unpublished", nil)
+	if err != nil {
+		log.Fatal(err)
+	}
+	req.Header.Set("api-key", DEVAPIKEY)
+	resp, err := client.Do(req)
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer resp.Body.Close()
+	var data interface{}
+	err = json.NewDecoder(resp.Body).Decode(&data)
+	if err != nil {
+		log.Fatal(err)
+	}
+	return data
+}
+
+func main() {
+	// Parse JSON
+	query, err := gojq.Parse(".[].id")
+	if err != nil {
+		log.Fatalln(err)
+	}
+	input := curl()
+	iter := query.Run(input) // or query.RunWithContext
+	for {
+		v, ok := iter.Next()
+		if !ok {
+			break
+		}
+		if err, ok := v.(error); ok {
+			log.Fatalln(err)
+		}
+		fmt.Printf("%1.0f\n", v)
+	}
+}
